// app/typing.d.ts
//@typescript-eslint/no-explicit-any
export type Car = {
    pricePerDay: ReactI18NextChildren | Iterable<ReactI18NextChildren>;
    tags: ReactI18NextChildren | Iterable<ReactI18NextChildren>;
    fuelCapacity: ReactI18NextChildren | Iterable<ReactI18NextChildren>;
    transmission: ReactI18NextChildren | Iterable<ReactI18NextChildren>;
    seatingCapacity: ReactI18NextChildren | Iterable<ReactI18NextChildren>;
    imageUrl: string | StaticImport;
    id: string;
    name: string;
    type: string;
    brand: string; // Added brand property
    title: string;
    description: string;
    image: string;

};